# examples/backend/common/openapi.yaml
openapi: 3.1.0
info:
  title: Backend Learning Admin API
  version: 1.0.0
  description: Shared contract implemented by NestJS, FastAPI, and Gin.
servers:
  - url: http://localhost:3001/api
    description: NestJS
  - url: http://localhost:3002/api
    description: FastAPI
  - url: http://localhost:3003/api
    description: Gin
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      operationId: refresh
      security: []
      responses:
        '200':
          description: Refresh token rotated; replay revokes the complete token family.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json: { schema: { $ref: '#/components/schemas/AccessToken' } }
        '401': { $ref: '#/components/responses/Problem' }
...
  /chat-runs:
    post:
      operationId: createChatRun
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ChatRunInput' } }
      responses:
        '202':
          description: Chat run accepted; events are streamed from the returned task.
          content:
            application/json: { schema: { $ref: '#/components/schemas/Task' } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
...
    Problem:
      description: Stable problem response.
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
...
  schemas:
    Uuid: { type: string, format: uuid }
...
    AccessToken:
      type: object
      additionalProperties: false
      required: [accessToken, expiresIn]
      properties:
        accessToken: { type: string }
        expiresIn: { type: integer, const: 900 }
...
    Task:
      type: object
      required: [id, type, status, progress, attempt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        type: { type: string }
        status: { type: string, enum: [queued, running, retrying, completed, failed, cancelled] }
        progress: { type: integer, minimum: 0, maximum: 100 }
        attempt: { type: integer, minimum: 0 }
        result: { type: [object, 'null'], additionalProperties: true }
...
    ChatRunInput:
      type: object
      required: [knowledgeBaseId, question]
      properties:
        knowledgeBaseId: { $ref: '#/components/schemas/Uuid' }
        question: { type: string, minLength: 1, maxLength: 8000 }
    FieldError:
      type: object
      required: [field, messages]
      properties:
        field: { type: string }
        messages: { type: array, items: { type: string } }
    Problem:
      type: object
      additionalProperties: false
      required: [status, code, detail, requestId]
      properties:
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string }
        detail: { type: string }
        requestId: { type: string }
        fields: { type: array, items: { $ref: '#/components/schemas/FieldError' } }